requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
CacheControl==0.12.10
cachetools==5.5.0
cachy==0.3.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.0
cleo==0.8.1
click==8.0.3
clikit==0.6.2
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.1
crashtest==0.3.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distlib==0.3.4
distro==1.7.0
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
docopt==0.6.2
exceptiongroup==1.2.2
filelock==3.6.0
fonttools==4.55.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
hdfs==2.7.3
html5lib==1.1
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jiter==0.7.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==8.10.0
msgpack==1.0.3
narwhals==1.13.5
netifaces==0.11.0
numpy==2.1.3
oauthlib==3.2.0
openai==1.54.4
packaging==21.3
pandas==2.2.3
pastel==0.2.1
pexpect==4.8.0
pillow==11.0.0
pkginfo==1.8.2
platformdirs==2.5.1
plotly==5.24.1
poetry==1.1.12
poetry-core==1.0.7
protobuf==5.28.3
psycopg2==2.9.10
ptyprocess==0.7.0
py4j==0.10.9.7
pyarrow==18.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylev==1.2.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pyspark==3.5.3
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2022.1
PyYAML==5.4.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==0.9.1
rich==13.9.4
SecretStorage==3.3.1
service-identity==18.1.0
shellingham==1.4.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.2
ssh-import-id==5.11
streamlit==1.40.1
systemd-python==234
tenacity==9.0.0
tiktoken==0.8.0
toml==0.10.2
tomlkit==0.9.2
tornado==6.4.1
tqdm==4.67.0
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
watchdog==6.0.0
webencodings==0.5.1
wordcloud==1.9.4
zipp==1.0.0
zope.interface==5.4.0
